:recycle: [Refactor] 랜덤 기사 코멘트 수정 & 파일명 수정
