+++ src/components/layout/GNB.tsx
@@ -73,7 +73,7 @@ const GNB = ({ userType = null }: GNBProps) => {
           <>
             <NavItem href="/request-quote">견적 요청</NavItem>
             <NavItem href="/find-mover">기사님 찾기</NavItem>
-            <NavItem href="/my-quotes">내 견적 관리</NavItem>
+            <NavItem href="/my-quote">내 견적 관리</NavItem>
           </>
         );
       default:
